Update insulin_glucose_pairing.ipynb
Change of the title of this notebook from "Identification of Critically-ill Patients Under Insulin Therapy for Glucose Control" to "Data extraction and Pairing of Insulin Inputs to Glucose Measurements in the ICU" being that describes better the main aim of this notebook. 
Addition of more authors.
Modification of the query that extracts the raw inputs of insulin. The STORETIME column is not imported. The INFXSTOP column don't have NULL values, it has 0 instead.
The plot "Time difference (CHARTTIME - STORETIME)" is not printed out anymore.
Addition of more sections: Differences between fingerstick and lab analyzer, Length of stay analysis, Glucose readings/ICU stays.
A new strategy to analyse the Non-adjusted datasets.