Update README.md
Title of the paper changed. This commit updates the title to the current title of the re-submitted version.
Before: Blood Glucose Management in the Intensive Care Unit: Curation Process and Insights from a Data-driven Approach.
Modified: Data-driven Curation Process for describing the Blood Glucose Management in the Intensive Care Unit.

Adding a brief description of the content in this folder.