Add import.sh
import.sh is a shell script that runs on any POSIX compliant shell.
It will import the MIMIC-III dataset and place the data into a single
SQLite database.

It follows the naming conventions of the pre-existing import.py, so the
output file is "mimic3.db" and tables are named in all lower-case from
the file name. For example, the data in "ADMISSIONS.csv" is saved in the
"admissions" table.

It works both with the demo dataset (.csv files) and the full dataset
(.csv.gz files) with no modification.

RAM and disk utilization is low, as it uses gzip to stream files
directly to the sqlite3 binary for loading. It also works quickly: on my
system it took 11 minutes for the full dataset (after a reboot and the
caches cleared) and under 5 seconds for the demo dataset.

The shell script was linted with "shellcheck" with no warnings/errors.